# [enhancement] proxy: change default config value and sql stmt of use.