@matrix-meow 2024-01-09
@volgariver6 Thanks for your contributions!

Title: [enhancement] proxy: change default config value and sql stmt of use.

Body: This pull request is an enhancement that changes the default configuration value and SQL statement for the "use" event in the proxy.

Changes:
- In `pkg/proxy/client_conn.go`, the `handleInitDB` function has been modified. The `redoStmts` slice is now appended with an `internalStmt` struct that contains the `cmdType` and `s` fields. The `s` field is no longer concatenated with the string "use ", but is now assigned the value of `e.db`.
- In `pkg/proxy/config.go`, the `defaultRebalanceInterval` variable has been changed from 30 seconds to 15 seconds.

Review:

1. In the `handleInitDB` function in `pkg/proxy/client_conn.go`, the change made to the `redoStmts` slice seems incorrect. The previous code appended an `internalStmt` struct with the `cmdType` set to `cmdInitDB` and the `s` field set to "use " concatenated with `e.db`. The new code simply assigns `e.db` to the `s` field. This change may result in incorrect behavior when handling the "use" event. It is recommended to revert this change and keep the concatenation of "use " and `e.db` in the `s` field.

2. In `pkg/proxy/config.go`, the change to `defaultRebalanceInterval` from 30 seconds to 15 seconds seems reasonable. However, it would be helpful to provide a comment explaining the reason for this change. Additionally, it would be beneficial to update the documentation to reflect this new default value.

Optimization:

1. The changes made in this pull request are relatively minor and do not require further optimization.

Security:

No security issues were identified in this pull request.